Catch errors in whitelisting
If the whitelists are missing or incorrectly parsed, the ‘ind_type’
will not be in the self.whitelist dict. This will result in a KeyError
exception. This catches that error and just passes rather than raising
it up and printing “Host” which is a very elusive error to track down.